formats-RPM: Move to separate repository. arm: Hi-MOV and block data transfer instructions. arm-elf: Correct nomenclature. cli: Fix default content parsing. formats-yaml: Better list formatting. formats-zip: Zip packer. fs: Better relative path generator, stream transfer. fs: DiskFile attribute caching optimization. fs: MemoryFile now has complete virtual file tree capabilities. gui: Various improvements and bug fixes. io-base: Some improvements here and there, bug fixes. io-serialization: Finished BinarySerializer. math: More matrix decomposition functions, higher FAtan resolution, cleanup. net: Fix FileDownloader. text: String join. util: Enum bitflags emulator, JVM context checker, ReflectionHash collection fixes. util: Primitive collection optimization.

+++ src/ctrmap/stdlib/io/serialization/SerializationIOStream.java
@@ -6,7 +6,7 @@ import java.io.IOException;
 
 public class SerializationIOStream extends DataIOStream {
 	
-	private int lastTracedSeek = -1;
+	private long lastTracedSeek = -1;
 	
 	public SerializationIOStream(IOStream ios) {
 		super(ios);
@@ -20,12 +20,12 @@ public class SerializationIOStream extends DataIOStream {
 		lastTracedSeek = -1;
 	}
 	
-	public int getMaxSeekSinceTrace() throws IOException {
+	public long getMaxSeekSinceTrace() throws IOException {
 		return Math.max(getPositionUnbased(), lastTracedSeek);
 	}
 	
 	@Override
-	public void seek(int pos) throws IOException {
+	public void seek(long pos) throws IOException {
 		lastTracedSeek = Math.max(lastTracedSeek, Math.max(getPositionUnbased(), pos));
 		super.seek(pos);
 	}
